refactor test scripts into single file
